Fix -Werror warnings across musica Fortran and C++ interfaces
- Replace Fortran allocate/free with C malloc/free in version string
  functions to fix Intel Debug allocator errors (thread-safe fix)
- Remove version_length param from InternalFree*Version signatures
- Fix -Wc-binding-type: use real(c_double) and integer(c_int) in
  BIND(C) procedures across tuvx and carma interface files
- Fix -Wconversion: use integer(c_size_t) for loop counters iterating
  over c_size_t bounds in tuvx and fortran/util.F90
- Fix -Wunused-parameter: remove unused musica_dk imports from pointer-
  return functions, fix solver_interface.hpp and v1_parse.cpp
- Suppress warnings from micm headers by adding SYSTEM includes
- Fix duplicate libmusica-fortran.a linker warning in example tests

+++ src/tuvx/interface_radiator_map.F90
@@ -115,7 +115,7 @@ end subroutine internal_delete_radiator_map
     class(radiator_t),          pointer :: f_radiator_ptr
     type(radiator_warehouse_t), pointer :: radiator_warehouse
     character(len=:),       allocatable :: f_radiator_name
-    integer                             :: i
+    integer(c_size_t)                   :: i
 
     ! result
     type(c_ptr) :: radiator_ptr
@@ -216,7 +216,7 @@ end subroutine internal_delete_radiator_map
 
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
 
-  integer function internal_get_number_of_radiators(radiator_map, error_code) &
+  integer(c_int) function internal_get_number_of_radiators(radiator_map, error_code) &
       bind(C, name="InternalGetNumberOfRadiators")
     use iso_c_binding, only: c_ptr, c_f_pointer, c_int
     use tuvx_radiator_warehouse, only: radiator_warehouse_t
@@ -257,7 +257,7 @@ end subroutine internal_delete_radiator_map
     ! variables
     type(radiator_warehouse_t), pointer :: f_radiator_warehouse
     character(len=:),       allocatable :: f_radiator_name
-    integer                             :: i
+    integer(c_size_t)                   :: i
 
     error_code = ERROR_NONE
     allocate(character(len=c_radiator_name_length) :: f_radiator_name)
@@ -297,7 +297,7 @@ end subroutine internal_delete_radiator_map
 
     ! variables
     type(radiator_warehouse_t), pointer :: f_radiator_warehouse
-    integer :: i, j
+    integer(c_size_t) :: i, j
     type(radiator_ptr), allocatable :: temp_radiators(:)
 
     error_code = ERROR_NONE
